simplified the discovery greatly by removing the endpoint since we know them already
