--- bindings/go/plugin/manager/plugin_implementations.go
@@ -1,497 +1,0 @@
-package manager
-
-import (
-	"bytes"
-	"context"
-	"encoding/base64"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"github.com/santhosh-tekuri/jsonschema/v6"
-	"log/slog"
-	"net/http"
-	"os"
-
-	descriptor "ocm.software/open-component-model/bindings/go/descriptor/runtime"
-	"ocm.software/open-component-model/bindings/go/runtime"
-)
-
-const (
-	XOCMRepositoryHeader = "X-Ocm-Repository"
-)
-
-// Endpoints
-const (
-	// UploadLocalResource defines the endpoint to upload a local resource to.
-	UploadLocalResource = "/local-resource/upload"
-	// DownloadLocalResource defines the endpoint to download a local resource.
-	DownloadLocalResource = "/local-resource/download"
-	// UploadResource defines the endpoint to upload a global resource to.
-	UploadResource = "/resource/upload"
-	// DownloadResource defines the endpoint to download a global resource.
-	DownloadResource = "/resource/download"
-	// UploadComponentVersion defines the endpoint to upload component versions to.
-	UploadComponentVersion = "/component-version/upload"
-	// DownloadComponentVersion defines the endpoint to download component versions.
-	DownloadComponentVersion = "/component-version/download"
-)
-
-func NewTypedRepositoryPlugin[T runtime.Typed](base *RepositoryPlugin) *TypedRepositoryPlugin[T] {
-	return &TypedRepositoryPlugin[T]{base}
-}
-
-type TypedRepositoryPlugin[T runtime.Typed] struct {
-	base *RepositoryPlugin
-}
-
-func (r *TypedRepositoryPlugin[T]) GetLocalResource(ctx context.Context, request GetLocalResourceRequest[T], credentials Attributes) error {
-	return r.base.GetLocalResource(ctx, GetLocalResourceRequest[runtime.Typed]{
-		Repository:     request.Repository,
-		Name:           request.Name,
-		Version:        request.Version,
-		Identity:       request.Identity,
-		TargetLocation: request.TargetLocation,
-	}, credentials)
-}
-
-func (r *TypedRepositoryPlugin[T]) AddLocalResource(ctx context.Context, request PostLocalResourceRequest[T], credentials Attributes) (*descriptor.Resource, error) {
-	return r.base.AddLocalResource(ctx, PostLocalResourceRequest[runtime.Typed]{
-		Repository:       request.Repository,
-		Name:             request.Name,
-		Version:          request.Version,
-		ResourceLocation: request.ResourceLocation,
-		Resource:         request.Resource,
-	}, credentials)
-}
-func (r *TypedRepositoryPlugin[T]) Ping(ctx context.Context) error {
-	return r.base.Ping(ctx)
-}
-
-func (r *TypedRepositoryPlugin[T]) AddComponentVersion(ctx context.Context, request PostComponentVersionRequest[T], credentials Attributes) error {
-	return r.base.AddComponentVersion(ctx, PostComponentVersionRequest[runtime.Typed]{
-		Repository: request.Repository,
-		Descriptor: request.Descriptor,
-	}, credentials)
-}
-
-func (r *TypedRepositoryPlugin[T]) GetComponentVersion(ctx context.Context, request GetComponentVersionRequest[T], credentials Attributes) (*descriptor.Descriptor, error) {
-	req := GetComponentVersionRequest[runtime.Typed]{
-		Name:       request.Name,
-		Version:    request.Version,
-		Repository: request.Repository,
-	}
-	return r.base.GetComponentVersion(ctx, req, credentials)
-}
-
-type RepositoryPlugin struct {
-	ID string
-
-	// config is used to start the plugin during a later phase.
-	config Config
-	path   string
-	client *http.Client
-	logger *slog.Logger
-
-	baseCtx context.Context
-
-	// capabilities we know which endpoint we will call, so check the schema for it.
-	endpoints []Endpoint
-}
-
-// This plugin implements all the given contracts.
-var (
-	_ ReadOCMRepositoryPluginContract[runtime.Typed]  = &RepositoryPlugin{}
-	_ WriteOCMRepositoryPluginContract[runtime.Typed] = &RepositoryPlugin{}
-	_ ReadResourcePluginContract                      = &RepositoryPlugin{}
-	_ WriteResourcePluginContract                     = &RepositoryPlugin{}
-)
-
-func NewRepositoryPlugin(baseCtx context.Context, logger *slog.Logger, client *http.Client, id string, path string, config Config, endpoints []Endpoint) *RepositoryPlugin {
-	return &RepositoryPlugin{
-		baseCtx:   baseCtx,
-		ID:        id,
-		path:      path,
-		config:    config,
-		logger:    logger,
-		client:    client,
-		endpoints: endpoints,
-	}
-}
-
-func (r *RepositoryPlugin) Ping(ctx context.Context) error {
-	r.logger.Info("Pinging plugin", "id", r.ID)
-
-	if err := call(ctx, r.client, "healthz", http.MethodGet); err != nil {
-		return fmt.Errorf("failed to ping plugin %s: %w", r.ID, err)
-	}
-
-	return nil
-}
-
-func (r *RepositoryPlugin) AddComponentVersion(ctx context.Context, request PostComponentVersionRequest[runtime.Typed], credentials Attributes) error {
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return err
-	}
-
-	for _, e := range r.endpoints {
-		if e.Location == UploadComponentVersion {
-			if len(e.Types) != 1 {
-				return fmt.Errorf("missing endpoint schema validation setup for AddComponentVersion; was len: %d", len(e.Types))
-			}
-
-			if err := r.validateEndpoint(request.Repository, e.Types[0].JSONSchema); err != nil {
-				return err
-			}
-		}
-	}
-
-	if err := call(ctx, r.client, UploadComponentVersion, http.MethodPost, WithPayload(request), WithHeader(credHeader)); err != nil {
-		return fmt.Errorf("failed to add component version with plugin %q: %w", r.ID, err)
-	}
-
-	return nil
-}
-
-func (r *RepositoryPlugin) GetComponentVersion(ctx context.Context, request GetComponentVersionRequest[runtime.Typed], credentials Attributes) (*descriptor.Descriptor, error) {
-	response := &descriptor.Descriptor{}
-	var params []KV
-	addParam := func(k, v string) {
-		params = append(params, KV{Key: k, Value: v})
-	}
-	addParam("name", request.Name)
-	addParam("version", request.Version)
-
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	repoHeader, err := toOCMRepoHeader(request.Repository)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, e := range r.endpoints {
-		if e.Location == DownloadComponentVersion {
-			if len(e.Types) != 1 {
-				return nil, fmt.Errorf("missing endpoint schema validation setup for AddComponentVersion; was len: %d", len(e.Types))
-			}
-
-			if err := r.validateEndpoint(request.Repository, e.Types[0].JSONSchema); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	if err := call(ctx, r.client, DownloadComponentVersion, http.MethodGet, WithResult(response), WithQueryParams(params), WithHeader(credHeader), WithHeader(repoHeader)); err != nil {
-		return nil, fmt.Errorf("failed to get component version %s:%s from %s: %w", request.Name, request.Version, r.ID, err)
-	}
-
-	return response, nil
-}
-
-func (r *RepositoryPlugin) AddLocalResource(ctx context.Context, request PostLocalResourceRequest[runtime.Typed], credentials Attributes) (*descriptor.Resource, error) {
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, e := range r.endpoints {
-		if e.Location == DownloadComponentVersion {
-			if len(e.Types) != 1 {
-				return nil, fmt.Errorf("missing endpoint schema validation setup for AddComponentVersion; was len: %d", len(e.Types))
-			}
-
-			if err := r.validateEndpoint(request.Repository, e.Types[0].JSONSchema); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	response := &descriptor.Resource{}
-	if err := call(ctx, r.client, UploadLocalResource, http.MethodPost, WithPayload(request), WithResult(response), WithHeader(credHeader)); err != nil {
-		return nil, fmt.Errorf("failed to add local resource %s: %w", r.ID, err)
-	}
-
-	return response, nil
-}
-
-func (r *RepositoryPlugin) GetLocalResource(ctx context.Context, request GetLocalResourceRequest[runtime.Typed], credentials Attributes) error {
-	var params []KV
-	addParam := func(k, v string) {
-		params = append(params, KV{Key: k, Value: v})
-	}
-	addParam("name", request.Name)
-	addParam("version", request.Version)
-	addParam("target_location_type", string(request.TargetLocation.LocationType))
-	addParam("target_location_value", request.TargetLocation.Value)
-	identityEncoded, err := json.Marshal(request.Identity)
-	if err != nil {
-		return err
-	}
-	identityBase64 := base64.StdEncoding.EncodeToString(identityEncoded)
-	addParam("identity", identityBase64)
-
-	repoHeader, err := toOCMRepoHeader(request.Repository) // Raw
-	if err != nil {
-		return err
-	}
-
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return err
-	}
-
-	//for _, e := range r.endpoints {
-	//	if e.Location == DownloadLocalResource {
-	//		if len(e.Types) != 1 {
-	//			return fmt.Errorf("missing endpoint schema validation setup for GetLocalResource; was len: %d", len(e.Types))
-	//		}
-	//
-	//		if err := r.validateEndpoint(request.Repository, e.Types[0].JSONSchema); err != nil {
-	//			return err
-	//		}
-	//	}
-	//}
-
-	if err := call(ctx, r.client, DownloadLocalResource, http.MethodGet, WithQueryParams(params), WithHeader(credHeader), WithHeader(repoHeader)); err != nil {
-		return fmt.Errorf("failed to get local resource %s:%s from %s: %w", request.Name, request.Version, r.ID, err)
-	}
-
-	_, err = os.Stat(request.TargetLocation.Value)
-	if err != nil {
-		return fmt.Errorf("failed to stat target file: %w", err)
-	}
-
-	return nil
-}
-
-func (r *RepositoryPlugin) AddGlobalResource(ctx context.Context, request PostResourceRequest, credentials Attributes) (*descriptor.Resource, error) {
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, e := range r.endpoints {
-		if e.Location == UploadResource {
-			if len(e.Types) != 1 {
-				return nil, fmt.Errorf("missing endpoint schema validation setup for GetLocalResource; was len: %d", len(e.Types))
-			}
-
-			if err := r.validateEndpoint(request.Resource.Access, e.Types[0].JSONSchema); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	response := &descriptor.Resource{}
-	if err := call(ctx, r.client, UploadResource, http.MethodPost, WithPayload(request), WithResult(response), WithHeader(credHeader)); err != nil {
-		return nil, fmt.Errorf("failed to add resource %s: %w", r.ID, err)
-	}
-
-	return response, nil
-}
-
-func (r *RepositoryPlugin) GetGlobalResource(ctx context.Context, request GetResourceRequest, credentials Attributes) error {
-	var params []KV
-	addParam := func(k, v string) {
-		params = append(params, KV{Key: k, Value: v})
-	}
-	addParam("target_location_type", string(request.TargetLocation.LocationType))
-	addParam("target_location_value", request.TargetLocation.Value)
-	resourceEncoded, err := json.Marshal(request.Resource)
-	if err != nil {
-		return err
-	}
-	resourceBase64 := base64.StdEncoding.EncodeToString(resourceEncoded)
-	addParam("resource", resourceBase64)
-
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return err
-	}
-
-	for _, e := range r.endpoints {
-		if e.Location == DownloadResource {
-			if len(e.Types) != 1 {
-				return fmt.Errorf("missing endpoint schema validation setup for GetGlobalResource; was len: %d", len(e.Types))
-			}
-
-			if err := r.validateEndpoint(request.Resource.Access, e.Types[0].JSONSchema); err != nil {
-				return err
-			}
-		}
-	}
-
-	if err := call(ctx, r.client, DownloadResource, http.MethodGet, WithQueryParams(params), WithHeader(credHeader)); err != nil {
-		return fmt.Errorf("failed to get local resource version %s: %w", r.ID, err)
-	}
-
-	_, err = os.Stat(request.TargetLocation.Value)
-	if err != nil {
-		return fmt.Errorf("failed to stat target file: %w", err)
-	}
-
-	return nil
-}
-
-// Call will use the plugin's constructed connection client to make a call to the specified
-// endpoint. The result will be marshalled into the provided response if not nil.
-func (r *RepositoryPlugin) Call(ctx context.Context, endpoint, method string, payload, response any, headers []KV, params []KV) error {
-	return call(
-		ctx,
-		r.client,
-		endpoint,
-		method,
-		WithPayload(payload),
-		WithResult(response),
-		WithHeaders(headers),
-		WithQueryParams(params),
-	)
-}
-
-func (r *RepositoryPlugin) SupportedRepositoryConfigTypes(ctx context.Context) ([]runtime.Type, error) {
-	var configTypes []runtime.Type
-
-	if err := call(ctx, r.client, "supported-credential-repository-config-types", http.MethodGet, WithResult(&configTypes)); err != nil {
-		return nil, fmt.Errorf("failed to retrieve identity for repository configuration %s: %w", r.ID, err)
-	}
-
-	return configTypes, nil
-}
-
-func (r *RepositoryPlugin) ConsumerIdentityForRepositoryConfig(ctx context.Context, config runtime.Typed) (runtime.Identity, error) {
-	var params []KV
-	addParam := func(k, v string) {
-		params = append(params, KV{Key: k, Value: v})
-	}
-	configEncoded, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	resourceBase64 := base64.StdEncoding.EncodeToString(configEncoded)
-	addParam("repository_config", resourceBase64)
-
-	identity := runtime.Identity{}
-
-	if err := call(ctx, r.client, "consumer-identity-for-repository-config", http.MethodGet, WithQueryParams(params), WithResult(&identity)); err != nil {
-		return nil, fmt.Errorf("failed to retrieve identity for repository configuration %s: %w", r.ID, err)
-	}
-
-	return identity, nil
-}
-
-func (r *RepositoryPlugin) Resolve(ctx context.Context, config runtime.Typed, identity runtime.Identity, credentials Attributes) (Attributes, error) {
-	var params []KV
-	addParam := func(k, v string) {
-		params = append(params, KV{Key: k, Value: v})
-	}
-
-	configEncoded, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-	resourceBase64 := base64.StdEncoding.EncodeToString(configEncoded)
-	addParam("repository_config", resourceBase64)
-
-	identityEncoded, err := json.Marshal(identity)
-	if err != nil {
-		return nil, err
-	}
-	identityBase64 := base64.StdEncoding.EncodeToString(identityEncoded)
-	addParam("identity", identityBase64)
-
-	credHeader, err := toCredentials(credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	resolved := Attributes{}
-
-	if err := call(ctx, r.client, "resolve-repository-credentials", http.MethodGet, WithQueryParams(params), WithResult(&resolved), WithHeader(credHeader)); err != nil {
-		return nil, fmt.Errorf("failed to retrieve identity for repository configuration %s: %w", r.ID, err)
-	}
-
-	return resolved, nil
-}
-
-func (r *RepositoryPlugin) validateEndpoint(obj runtime.Typed, jsonSchema []byte) error {
-	valid, err := validatePlugin(obj, jsonSchema)
-	if err != nil {
-		return fmt.Errorf("failed to validate plugin %q: %w", r.ID, err)
-	}
-	if !valid {
-		return fmt.Errorf("validation of plugin %q failed for get local resource", r.ID)
-	}
-
-	return nil
-}
-
-func toCredentials(credentials Attributes) (KV, error) {
-	rawCreds, err := json.Marshal(credentials)
-	if err != nil {
-		return KV{}, err
-	}
-	return KV{
-		Key:   "Authorization",
-		Value: string(rawCreds),
-	}, nil
-}
-
-func toOCMRepoHeader(repository runtime.Typed) (KV, error) {
-	raw, err := json.Marshal(repository)
-	if err != nil {
-		return KV{}, err
-	}
-	return KV{
-		Key:   XOCMRepositoryHeader,
-		Value: string(raw),
-	}, nil
-}
-
-func validatePlugin(typ runtime.Typed, jsonSchema []byte) (bool, error) {
-	c := jsonschema.NewCompiler()
-	unmarshaler, err := jsonschema.UnmarshalJSON(bytes.NewReader(jsonSchema))
-	var v any
-	if err := json.Unmarshal(jsonSchema, &v); err != nil {
-		return true, err
-	}
-
-	if err := c.AddResource("schema.json", unmarshaler); err != nil {
-		return true, fmt.Errorf("failed to add schema.json: %w", err)
-	}
-	sch, err := c.Compile("schema.json")
-	if err != nil {
-		return true, fmt.Errorf("failed to compile schema.json: %w", err)
-	}
-
-	// need to marshal the interface into a JSON format.
-	content, err := json.Marshal(typ)
-	if err != nil {
-		return true, fmt.Errorf("failed to marshal type: %w", err)
-	}
-	// once marshalled, we create a map[string]any representation of the marshaled content.
-	unmarshalledType, err := jsonschema.UnmarshalJSON(bytes.NewReader(content))
-	if err != nil {
-		return true, fmt.Errorf("failed to unmarshal : %w", err)
-	}
-
-	if _, ok := unmarshalledType.(string); ok {
-		// TODO: In _not_ POC this should be either a type switch, or some kind of exclusion or we should change how
-		// we register and look up plugins to avoid validating when listing or for certain plugins.
-		// skip validation if the passed in type is of type string.
-		return true, nil
-	}
-
-	// finally, validate map[string]any against the loaded schema
-	if err := sch.Validate(unmarshalledType); err != nil {
-		var typRaw bytes.Buffer
-		err = errors.Join(err, json.Indent(&typRaw, content, "", "  "))
-		var schemaRaw bytes.Buffer
-		err = errors.Join(err, json.Indent(&schemaRaw, jsonSchema, "", "  "))
-		return true, fmt.Errorf("failed to validate schema for\n%s\n---SCHEMA---\n%s\n: %w", typRaw.String(), schemaRaw.String(), err)
-	}
-
-	return true, nil
-}
